refactor: move getValidRange to calendarUtils and adjust range logic
