+++ frontend/src/app/helper/utils/calendarUtils.tsx
@@ -75,8 +75,10 @@ export const createRenderEventContent = (userType: UserType) => {
 };
 
 export const getValidRange = (createdAt: string) => {
+  // The valid range starts from the account creation date and ends at the end of the month, three months later.
   const now = new Date();
-  const end = new Date(now.getFullYear(), now.getMonth() + 3, 0);
+  const end = new Date(now.getFullYear(), now.getMonth() + 4, 1);
+
   return {
     start: createdAt.split("T")[0],
     end: end.toISOString().split("T")[0],
@@ -103,4 +105,43 @@ export const createDayCellDidMount = (
   };
 
   return DayCellDidMount;
+};
+
+export const getClassSlotTimesForCalendar = () => {
+  // Step 1: Get the user's timezone offset from UTC in minutes
+  const timezoneOffsetMinutes = new Date().getTimezoneOffset();
+
+  // Step 2: Calculate class start and end times, which are UTC midnight and noon (in Japan time: start at 09:00 and end at 21:00)
+  const utcClassesStartTime = new Date(Date.UTC(1970, 0, 1, 0, 0, 0)); // 00:00 UTC
+  const utcClassesEndTime = new Date(Date.UTC(1970, 0, 1, 12, 0, 0)); // 12:00 UTC
+
+  // Step 3: Adjust UTC times to the user's local time by subtracting the timezone offset
+  const localClassesStartTime = new Date(
+    utcClassesStartTime.getTime() - timezoneOffsetMinutes * 60 * 1000,
+  );
+  const localClassesEndTime = new Date(
+    utcClassesEndTime.getTime() - timezoneOffsetMinutes * 60 * 1000,
+  );
+
+  // Step 4: Format times to "HH:mm:ss"
+  const formatTime = (date: Date) => date.toISOString().substring(11, 19);
+  const formattedLocalClassesStartTime = formatTime(localClassesStartTime);
+  const formattedLocalClassesEndTime = formatTime(localClassesEndTime);
+
+  // Step 5: Compare and set correct min and max times
+  // In some time zones (e.g., Vancouver), the local class start time (17:00) is later than the end time (05:00).
+  // In such cases, the min and max times need to be reversed.
+  const slotMinTime =
+    formattedLocalClassesStartTime < formattedLocalClassesEndTime
+      ? formattedLocalClassesStartTime
+      : formattedLocalClassesEndTime;
+  const slotMaxTime =
+    formattedLocalClassesEndTime > formattedLocalClassesStartTime
+      ? formattedLocalClassesEndTime
+      : formattedLocalClassesStartTime;
+
+  return {
+    slotMinTime,
+    slotMaxTime,
+  };
 };
